refactor(package): rename packages
Rename main package `com.nnk.springboot` to `com.pcs.tradingapp`.
Rename subpackages accordingly.
