--- Poseiden-skeleton/src/main/java/com/nnk/springboot/Application.java
@@ -1,12 +1,0 @@
-package com.nnk.springboot;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Application {
-
-	public static void main(String[] args) {
-		SpringApplication.run(Application.class, args);
-	}
-}
