--- Poseiden-skeleton/src/main/java/com/nnk/springboot/repositories/BidListRepository.java
@@ -1,9 +1,0 @@
-package com.nnk.springboot.repositories;
-
-import com.nnk.springboot.domain.BidList;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-
-public interface BidListRepository extends JpaRepository<BidList, Integer> {
-
-}
