--- Poseiden-skeleton/src/main/java/com/nnk/springboot/repositories/RatingRepository.java
@@ -1,8 +1,0 @@
-package com.nnk.springboot.repositories;
-
-import com.nnk.springboot.domain.Rating;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface RatingRepository extends JpaRepository<Rating, Integer> {
-
-}
