--- Poseiden-skeleton/src/main/java/com/nnk/springboot/repositories/RuleNameRepository.java
@@ -1,8 +1,0 @@
-package com.nnk.springboot.repositories;
-
-import com.nnk.springboot.domain.RuleName;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-
-public interface RuleNameRepository extends JpaRepository<RuleName, Integer> {
-}
