--- Poseiden-skeleton/src/main/java/com/nnk/springboot/repositories/TradeRepository.java
@@ -1,8 +1,0 @@
-package com.nnk.springboot.repositories;
-
-import com.nnk.springboot.domain.Trade;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-
-public interface TradeRepository extends JpaRepository<Trade, Integer> {
-}
