--- Poseiden-skeleton/src/main/java/com/nnk/springboot/repositories/UserRepository.java
@@ -1,12 +1,0 @@
-package com.nnk.springboot.repositories;
-
-import com.nnk.springboot.domain.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
-
-
-public interface UserRepository extends JpaRepository<User, Integer>, JpaSpecificationExecutor<User> {
-
-}
