+++ Poseiden-skeleton/src/main/java/com/pcs/tradingapp/repositories/BidListRepository.java
@@ -1,0 +1,10 @@
+package com.pcs.tradingapp.repositories;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.pcs.tradingapp.domain.BidList;
+
+
+public interface BidListRepository extends JpaRepository<BidList, Integer> {
+
+}
