+++ Poseiden-skeleton/src/main/java/com/pcs/tradingapp/repositories/RatingRepository.java
@@ -1,0 +1,9 @@
+package com.pcs.tradingapp.repositories;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.pcs.tradingapp.domain.Rating;
+
+public interface RatingRepository extends JpaRepository<Rating, Integer> {
+
+}
