+++ Poseiden-skeleton/src/main/java/com/pcs/tradingapp/repositories/UserRepository.java
@@ -1,0 +1,11 @@
+package com.pcs.tradingapp.repositories;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
+
+import com.pcs.tradingapp.domain.User;
+
+
+public interface UserRepository extends JpaRepository<User, Integer>, JpaSpecificationExecutor<User> {
+
+}
